ci: improve caching of dependencies
Our project is very heavy on dependencies, it takes ~10 minutes to fetch
and build them all. To keep our CI builds fast, we need to cache as much
as we can.

In this pull request, I am reworking the CI structure so that we always
start with building the project for the given platform, and then
fetching the artefacts produced by that build to speed up all subsequent
steps like testng and linting

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>