ci: add `--target` to `cargo test`
I noticed in GH Action output that `cargo test` is compiling our
dependencies again. Maybe it is not correctly reusing the cache?

In this commit, I am adding `--target` flag to `cargo test`, using
the same value that was provided to `cargo build` which is producing
the build artifacts saved to the cache.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>