ci: improve caching of build artifacts (#41)
* ci: add `--target` to `cargo test`

I noticed in GH Action output that `cargo test` is compiling our
dependencies again. Maybe it is not correctly reusing the cache?

In this commit, I am adding `--target` flag to `cargo test`, using
the same value that was provided to `cargo build` which is producing
the build artifacts saved to the cache.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>

* ci: build all targets (bins, libs, tests, etc.)

I noticed in GH Action output that `cargo test` is compiling our
dependencies again. Maybe the build step did not compile everything
that's needed by our tests?

In this commit, I am adding `--all-targets` flag to `cargo build`,
hoping that the subsequent `cargo test` will not need to compile
anything.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>

---------

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>