feat: print initial job count on `zinniad` start (#230)
Announce the total number of completed jobs immediately at startup.

Before this change, we would not announce it until the first job was
completed. That led to suboptimal UX: the Station initially displayed
job count 0. After a while, the counter jumped to a much higher value.

With this change in place, Filecoin Station should have the right value
of our job counter as soon as we load it from our persisted state.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>
Co-authored-by: Julian Gruber <julian@juliangruber.com>